+++ src/components/spa/classification.tsx
@@ -132,7 +132,6 @@ const Classification = ({ autoProcId, type = "2d" }: ClassificationProps) => {
           <Select
             isDisabled={!data || data.total < 1}
             aria-labelledby='sortlabel'
-            bg='white'
             onChange={(e) => setSortType(e.target.value as SortTypes)}
             size='xs'
             w='180px'
@@ -148,7 +147,7 @@ const Classification = ({ autoProcId, type = "2d" }: ClassificationProps) => {
             startFrom='start'
             page={classPage}
             onChange={setClassPage}
-            total={pageAmount}
+            total={!data || data.total < 1 ? 1 : pageAmount}
           />
         </HStack>
       </Stack>